INC-95 fix: payments added to the settings tab
